requirements.txt
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
ConfigArgParse==1.5.3
Flask==2.1.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.9.0
MarkupSafe==2.1.1
msgpack==1.0.3
paho-mqtt==1.6.1
psutil==5.9.1
pycparser==2.21
pywin32==304
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
